requirements.txt
Spire.Presentation
pypdf
